+++ src/bunit.web/ComponentFactories/StubComponentFactory.cs
@@ -22,7 +22,7 @@ internal sealed class StubComponentFactory : IComponentFactory
 	public IComponent Create(Type componentType)
 	{
 		var typeToCreate = StubType.MakeGenericType(componentType);
-		return (IComponent)Activator.CreateInstance(typeToCreate, new object?[] { replacementContent })!;
+		return (IComponent)Activator.CreateInstance(typeToCreate, replacementContent)!;
 	}
 }
 #endif